Improve the bxgenbb_inspector application
The bxgenbb_inspector application is now able to save randomized
primary events in files, in addition to the original histogramming
output.
Two modes are available:
- 'plain' : Events are serialized in one unique file as
  plain genbb::primary_event objects.
- 'bank' : Events are serialized in files, using a dedicated
  bank in some datatools::things event records. This mode
  enables to reload the event using a genbb::from_file generator.

+++ source/bxgenbb_help/examples/ex01/process.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env bash
 
+clean=1
+
+if [ "x$1" = "x--no-clean" ]; then
+    clean=0
+    shift 1
+fi
+
 which bxquery > /dev/null 2>&1
 if [ $? -ne 0 ]; then
     echo "ERROR: Missing Bayeux's bxquery !" >&2
@@ -65,22 +72,60 @@ if [ $? -ne 0 ]; then
     my_exit 1
 fi
 
-echo "Shoot events..." 1>&2
+echo "Shoot Co60 events..." 1>&2
 ${genbb_inspector_bin} \
     --configuration config/manager.conf \
     --action shoot \
     --generator "Co60" \
     --prng-seed 314159 \
-    --number-of-events 55 \
+    --number-of-events 100 \
+    --output-mode "histograms" \
     --output-file "histos_Co60.root"
 if [ $? -ne 0 ]; then
     echo "ERROR: Shoot events failed !" 1>&2
     my_exit 1
 fi
 
-rm -f ./ex01
-rm -f ./histos_Co60.root
-rm -fr ${build_dir}
+echo "" 1>&2
+echo "Shoot only one bb2nu_Se82_HE event..." 1>&2
+${genbb_inspector_bin} \
+    --configuration config/manager.conf \
+    --action shoot \
+    --generator "bb2nu_Se82_HE" \
+    --prng-seed 314159 \
+    --number-of-events 1 \
+    --output-mode "plain" \
+    --output-file "histos_bb2nu_Se82_HE_one_shot.xml"
+if [ $? -ne 0 ]; then
+    echo "ERROR: One shoot event failed !" 1>&2
+    my_exit 1
+fi
+
+echo "" 1>&2
+echo "Shoot many bb2nu_Se82_HE events..." 1>&2
+${genbb_inspector_bin} \
+    --configuration config/manager.conf \
+    --action shoot \
+    --generator "bb2nu_Se82_HE" \
+    --prng-seed 314159 \
+    --number-of-events 27 \
+    --output-mode "bank" \
+    --output-bank-max-records 10 \
+    --output-file "histos_bb2nu_Se82_HE_0.xml" \
+    --output-file "histos_bb2nu_Se82_HE_1.xml" \
+    --output-file "histos_bb2nu_Se82_HE_2.xml"
+if [ $? -ne 0 ]; then
+    echo "ERROR: Shoot events failed !" 1>&2
+    my_exit 1
+fi
+
+if [ $clean -eq 1 ]; then
+    rm -f ./ex01
+    rm -f ./histos_Co60.root
+    rm -f ./histos_bb2nu_Se82_HE_one_shot.xml
+    rm -f ./histos_bb2nu_Se82_HE_?.xml
+    rm -fr ${build_dir}
+fi
 
 cd ${opwd}
